refactor: filter, posts, educations의 type 설정 추가
