+++ src/type/filter.ts
@@ -45,4 +45,4 @@ export interface IMaxPrice {
 
 export type TStatus = "수강신청중" | "수강신청예정" | "마감" | "전체";
 export type TDate = string;
-export type TPrice = number;
+export type TPrice = string;
